Significant refactoring and cleanup
From now on, top-level directories in src/include/darma should
be self-contained, except for a dependence on src/include/darma/utility
(which, in turn, should not depend on anything else in src/include/darma).
Any parts that are dependent on another top-level folder should be
in an explicit subfolder stating such; for instance when we move key to
the top level, we'll want to have a subdirectory named
src/include/darma/key/dependent_on/serialization that contains a file
named key_serialization.h, which has the serializers for the key
stuff.

We've started this process with keyword_arguments and utility; all tests
pass.